# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TransactionCreate": {
        "additionalProperties": false,
        "description": "A ledger posting. See `app/modules/transactions/rules.py` for the rules per type. Amounts\nare Money objects in the posting `currency`; quantities and rates are strings.",
        "properties": {
          "asset_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "currency": {
            "description": "Currency of the posting",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "fees": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "fx_rate_as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the rate was valid (with a time zone)",
            "title": "Fx Rate As Of"
          },
          "fx_rate_source": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fx Rate Source"
          },
          "fx_rate_to_portfolio_currency": {
            "anyOf": [
              {
                "examples": [
                  "1234.50"
                ],
                "pattern": "^-?\\d{1,20}(\\.\\d{1,12})?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "1 unit of the posting currency in portfolio currency units. Required when the posting currency differs from the portfolio's.",
            "title": "Fx Rate To Portfolio Currency"
          },
          "gross_amount": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ],
            "description": "For BUY/SELL derived from quantity x unit price when omitted"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "quantity": {
            "anyOf": [
              {
                "examples": [
                  "1234.50"
                ],
                "pattern": "^-?\\d{1,20}(\\.\\d{1,12})?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Units, for trades and transfers",
            "title": "Quantity"
          },
          "settlement_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settlement Date"
          },
          "taxes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "trade_date": {
            "format": "date",
            "title": "Trade Date",
            "type": "string"
          },
          "transaction_type": {
            "enum": [
              "BUY",
              "SELL",
              "DEPOSIT",
              "WITHDRAWAL",
              "DIVIDEND",
              "INTEREST",
              "COUPON",
              "RENT",
              "DISTRIBUTION",
              "FEE",
              "TAX",
              "TRANSFER_IN",
              "TRANSFER_OUT",
              "VALUATION_ADJUSTMENT"
            ],
            "title": "Transaction Type",
            "type": "string"
          },
          "unit_price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "transaction_type",
          "trade_date",
          "currency"
        ],
        "title": "TransactionCreate",
        "type": "object"
      },
...
      "TransactionPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TransactionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` for the next page; null at the end",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "TransactionPage",
        "type": "object"
      },
...
      "ValuationPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ValuationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` for the next page; null at the end",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ValuationPage",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "OIDC access token from the WealthSphere realm",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "WealthSphere",
      "url": "https://github.com/kmdmtisya/KMD-Sphere-V2"
    },
    "description": "Backend for the WealthSphere mobile apps. The backend is authoritative for all financial calculations. Money, quantities and rates are JSON strings. Errors use RFC 7807 problem+json and every response carries X-Correlation-ID.",
    "summary": "Portfolio and wealth intelligence API",
    "title": "WealthSphere API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/portfolios/{portfolio_id}/transactions": {
      "get": {
        "description": "The ledger, newest first, with optional filters and cursor pagination.",
        "operationId": "list_transactions_api_v1_portfolios__portfolio_id__transactions_get",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Portfolio Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "transaction_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "BUY",
                    "SELL",
                    "DEPOSIT",
                    "WITHDRAWAL",
                    "DIVIDEND",
                    "INTEREST",
                    "COUPON",
                    "RENT",
                    "DISTRIBUTION",
                    "FEE",
                    "TAX",
                    "TRANSFER_IN",
                    "TRANSFER_OUT",
                    "VALUATION_ADJUSTMENT"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Transaction Type"
            }
          },
          {
            "in": "query",
            "name": "asset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          },
          {
            "in": "query",
            "name": "from_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "in": "query",
            "name": "to_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "No such portfolio for this user"
          },
          "422": {
            "description": "Invalid filter or cursor"
          },
          "429": {
            "description": "Rate limit exceeded (docs/security.md, API protection)",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Transactions",
        "tags": [
          "transactions"
        ]
      },
      "post": {
        "description": "Posts an entry to the portfolio's ledger. Posted entries never change.",
        "operationId": "post_transaction_api_v1_portfolios__portfolio_id__transactions_post",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Portfolio Id",
              "type": "string"
            }
          },
          {
            "description": "Client-chosen key (1-128 visible ASCII). A retry with the same key and content returns the original entry (200, Idempotent-Replayed: true).",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[\\x21-\\x7e]{1,128}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client-chosen key (1-128 visible ASCII). A retry with the same key and content returns the original entry (200, Idempotent-Replayed: true).",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Idempotent replay of an earlier posting",
            "headers": {
              "Idempotent-Replayed": {
                "description": "true when the response repeats an earlier posting with this key",
                "schema": {
                  "enum": [
                    "true"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "No such portfolio for this user"
          },
          "409": {
            "description": "Archived portfolio, quantity above the units held, or an Idempotency-Key reused for a different entry"
          },
          "422": {
            "description": "Invalid entry"
          },
          "429": {
            "description": "Rate limit exceeded (docs/security.md, API protection)",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Post Transaction",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/portfolios/{portfolio_id}/transactions/{transaction_id}": {
      "get": {
        "description": "One ledger entry, with its reversal link.",
        "operationId": "get_transaction_api_v1_portfolios__portfolio_id__transactions__transaction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Portfolio Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Transaction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "No such transaction for this user"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Rate limit exceeded (docs/security.md, API protection)",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Transaction",
        "tags": [
          "transactions"
        ]
      }
    },
...
    "/api/v1/portfolios/{portfolio_id}/valuations": {
      "get": {
        "description": "Valuations, newest as-of first.",
        "operationId": "list_valuations_api_v1_portfolios__portfolio_id__valuations_get",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Portfolio Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "asset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValuationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "No such portfolio for this user"
          },
          "422": {
            "description": "Invalid cursor"
          },
          "429": {
            "description": "Rate limit exceeded (docs/security.md, API protection)",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Valuations",
        "tags": [
          "valuations"
        ]
      },
      "post": {
        "description": "Records the value of an asset at a moment, with where the value comes from.",
        "operationId": "create_valuation_api_v1_portfolios__portfolio_id__valuations_post",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Portfolio Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValuationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValuationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "No such portfolio for this user"
          },
          "409": {
            "description": "A valuation already exists at that moment, or the portfolio is archived"
          },
          "422": {
            "description": "Invalid valuation"
          },
          "429": {
            "description": "Rate limit exceeded (docs/security.md, API protection)",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Valuation",
        "tags": [
          "valuations"
        ]
      }
    },
    "/api/v1/portfolios/{portfolio_id}/valuations/latest": {
      "get": {
        "description": "For each asset, the valuation with the latest as-of at or before `at`.",
        "operationId": "latest_valuations_api_v1_portfolios__portfolio_id__valuations_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Portfolio Id",
              "type": "string"
            }
          },
          {
            "description": "Values as of this moment (default: now)",
            "in": "query",
            "name": "at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Values as of this moment (default: now)",
              "title": "At"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ValuationOut"
                  },
                  "title": "Response Latest Valuations Api V1 Portfolios  Portfolio Id  Valuations Latest Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "No such portfolio for this user"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Rate limit exceeded (docs/security.md, API protection)",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Latest Valuations",
        "tags": [
          "valuations"
        ]
      }
    },
    "/api/v1/portfolios/{portfolio_id}/valuations/{valuation_id}": {
      "delete": {
        "description": "Removes a mistaken valuation (audited).",
        "operationId": "delete_valuation_api_v1_portfolios__portfolio_id__valuations__valuation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Portfolio Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "valuation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Valuation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "No such valuation for this user"
          },
          "409": {
            "description": "The portfolio is archived"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Rate limit exceeded (docs/security.md, API protection)",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Valuation",
        "tags": [
          "valuations"
        ]
      },
      "get": {
        "description": "One valuation.",
        "operationId": "get_valuation_api_v1_portfolios__portfolio_id__valuations__valuation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Portfolio Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "valuation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Valuation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValuationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "No such valuation for this user"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Rate limit exceeded (docs/security.md, API protection)",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Valuation",
        "tags": [
          "valuations"
        ]
      },
      "patch": {
        "description": "Corrects the value, source or note (audited). The as-of time cannot change.",
        "operationId": "update_valuation_api_v1_portfolios__portfolio_id__valuations__valuation_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Portfolio Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "valuation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Valuation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValuationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValuationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid or expired access token"
          },
          "404": {
            "description": "No such valuation for this user"
          },
          "409": {
            "description": "The portfolio is archived"
          },
          "422": {
            "description": "Invalid change"
          },
          "429": {
            "description": "Rate limit exceeded (docs/security.md, API protection)",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Valuation",
        "tags": [
          "valuations"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development",
      "url": "http://127.0.0.1:8000"
    }
  ],
...
}